+++ src-inst/flow_storm/api.clj
@@ -434,9 +434,10 @@
   `(instrument* {:tracing-disabled? true} ~form))
 
 (defn- read-rtrace-tag* [config form]
-  (let [full-config (merge config
-                           (meta form))]
-    `(runi ~full-config ~form)))
+  (let [{:keys [flow-id] :as full-config} (merge config (meta form))]
+    `(do
+       (runi ~full-config ~form)
+       (dbg-api/jump-to-last-expression-in-this-thread ~flow-id))))
 
 (defn read-rtrace-tag [form]  (read-rtrace-tag* {:flow-id 0} form))
 (defn read-rtrace0-tag [form] (read-rtrace-tag* {:flow-id 0} form))