+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 	
 ### Changes
 	
+	- Made search functionality faster and simplify it's code
+	
 ### Bugs fixed
 
-	- Don't crash if tracer functions are called before the system is fully started
+	- Don't crash if tracer functions are called before the system is fully started    
 
 ## 3.4.1 (17-04-2023)