+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 	- Add "Highlight current frame" on the call tree tool
 	- Add stack tab to code stepping tool
 	- Add a result pane on the functions list tool
+	- Add double click on calls tree tool node steps code
 	
 ### Changes