+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Changes
 	
 	- Made search functionality faster and simplify it's code
+	- #rtrace automatically opens the code stepping tool in the last position
 	
 ### Bugs fixed