+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 	
 	- Made search functionality faster and simplify it's code
 	- #rtrace automatically opens the code stepping tool in the last position
+	- Upgrade hansel to 0.1.54
 	
 ### Bugs fixed